Change filesystem encryption to take advantage of previous commit

+++ PC/secure-pc/filesystem/encryption.py
@@ -1,6 +1,7 @@
 import os
 import shutil
 import logging
+from typing import List
 
 from .constants import BACKUP_DIR
 from .tolerance import *
@@ -8,19 +9,19 @@ from ..security.encryption import AES256Encryption, Key
 
 _logger = logging.getLogger('file_encryption')
 
-def encrypt_file(path: str, key: Key, iv: bytes = None) -> bytes:
+def encrypt_file(path: str, key: Key) -> None:
     """
-    Encrypts a file using AES-256 CTR mode
+    Encrypts a file using AES-256 EAX mode.
+    The IV will be stored in the first 16 bytes of the file.
+
     :param path: the path to the file to encrypt
     :param key: the key to encrypt the file with
-    :param iv: the iv to encrypt the file with (if None, one will be generated)
-    :return: the iv used in the encryption
     """
     with open(path, 'rb') as f:
         data = f.read()
 
     cipher = AES256Encryption(key, AES256Encryption.MODE_EAX)
-    encrypted = cipher.encrypt(data, iv=iv)
+    encrypted = cipher.encrypt(data)
     log_encryption_start(path)
 
     if not os.path.isdir(BACKUP_DIR):
@@ -33,21 +34,20 @@ def encrypt_file(path: str, key: Key, iv: bytes = None) -> bytes:
     os.remove(os.path.join(BACKUP_DIR, os.path.basename(path)))
     log_encryption_end(path)
 
-    return cipher.iv
 
-
-def decrypt_file(path: str, key: Key, iv: bytes) -> None:
+def decrypt_file(path: str, key: Key) -> None:
     """
-    Decrypts a file using AES-256 CTR mode
+    Decrypts a file using AES-256 EAX mode.
+    The IV is assumed to be on the first 16 bytes of the file.
+
     :param path: the path to the file to decrypt
     :param key: the key to decrypt the file with
-    :param iv: the iv that was used to encrypt the file with
     """
     with open(path, 'rb') as f:
         data = f.read()
 
     cipher = AES256Encryption(key, AES256Encryption.MODE_EAX)
-    decrypted = cipher.decrypt(data, iv=iv)
+    decrypted = cipher.decrypt(data)
     log_decryption_start(path)
 
     if not os.path.isdir(BACKUP_DIR):
@@ -61,7 +61,7 @@ def decrypt_file(path: str, key: Key, iv: bytes) -> None:
     log_decryption_end(path)
 
 
-def encrypt_all(files: Dict[str, bytes], key: Key) -> None:
+def encrypt_all(files: List[str], key: Key) -> None:
     """
     Encrypts all files that need to be encrypted, even those
     that may have remained decrypted and are logged.
@@ -77,27 +77,27 @@ def encrypt_all(files: Dict[str, bytes], key: Key) -> None:
             if os.path.isfile(backup):
                 os.remove(backup)  # We don't want to have a backup we don't need
             continue
-        if status in ('encrypting', 'decrypting'):
-            if os.path.isfile(backup):
-                _logger.debug('Restoring {} from backup'.format(path))
-                shutil.copy2(backup, path)
-                status = 'decrypted'
+        if status in ('encrypting', 'decrypting') and os.path.isfile(backup):
+            _logger.debug('Restoring {} from backup'.format(path))
+            shutil.copy2(backup, path)
+            if status == 'encrypting':
+                status = 'decrypted' # in this case, the restored backup will be decrypted
         if status == 'decrypted':
-            encrypt_file(path, key, files[path])
+            encrypt_file(path, key)
 
-    for path, iv in files.items():
+    for path in files:
         if path in file_status:
             continue  # We've already handled it
-        encrypt_file(path, key, iv)
+        encrypt_file(path, key)
     clear_log()
 
 
-def decrypt_all(files: Dict[str, bytes], key: Key) -> None:
+def decrypt_all(files: List[str], key: Key) -> None:
     """
-    Decrypts all files.
+    Decrypts all files given.
 
-    :param files: dictionary mapping file paths to their encryption IVs
+    :param files: list of paths to the files to decrypt
     :param key: the key to decrypt the files with
     """
-    for path, iv in files.items():
-        decrypt_file(path, key, iv)
+    for path in files:
+        decrypt_file(path, key)
